# Adding indexes to tables used by LogPoller to speed up the execution