Some of the queries used by LogPoller are not hitting indexes and generating increased CPU load on the database to perform sequential scans or in-memory ordering

Queries and indexes were tested on my local machine with ~30kk rows in evm_log and ~30kk rows in evm_log_poller_block
<details>
<summary>Test data</summary>

```sql
CREATE EXTENSION pgcrypto;

insert into evm_chains (id, cfg, created_at, updated_at, enabled)
VALUES (1, '{}', now(), now(), true),
       (2, '{}', now(), now(), true),
       (3, '{}', now(), now(), true);

insert into evm_log_poller_blocks
select floor(random() * 3 + 1)::int, gen_random_bytes(32), generate_series(1, 10000000), now(), now();

INSERT INTO evm_logs
(evm_chain_id, log_index, block_hash, block_number, address, event_sig, topics, tx_hash, data, created_at,
 block_timestamp)
SELECT floor(random() * 3 + 1)::int                                                  as evm_chain_id,
       generate_series(1, 1000000)                                                   as log_index,
       gen_random_bytes(32)                                                           as block_hash,
       generate_series(1000001, 2000000)                                            as block_number,
       gen_random_bytes(32)                                                           as address,
       gen_random_bytes(32)                                                           as event_sig,
       ARRAY [gen_random_bytes(32) ]                                                  as topics,
       gen_random_bytes(32)                                                           as tx_hash,
       gen_random_bytes(32)                                                           as data,
       timestamp '2023-01-01 00:00:00' +
       random() * (timestamp '2023-12-31 23:59:59' - timestamp '2023-01-01 00:00:00') as created_at,
       timestamp '2023-01-01 00:00:00' +
       random() * (timestamp '2023-12-31 23:59:59' - timestamp '2023-01-01 00:00:00') as block_timestamp;
```


</details>



### Index added to the evm_log_poller_blocks
<details>
   <summary> Explain analyze test -  900ms -> 0.3ms </summary>
Before

```
----------
 Limit  (cost=0.56..0.77 rows=1 width=62) (actual time=900.497..900.498 rows=1 loops=1)
   ->  Index Scan Backward using log_poller_blocks_pkey on evm_log_poller_blocks  (cost=0.56..2073416.10 rows=10006027 width=62) (actual time=900.494..900.494 rows=1 loops=1)
         Index Cond: (evm_chain_id = '420'::numeric)
 Planning Time: 0.238 ms
 Execution Time: 900.531 ms
```

After

```
Limit  (cost=0.56..0.72 rows=1 width=62) (actual time=0.165..0.166 rows=1 loops=1)
   ->  Index Scan using idx_evm_chain_id_block_number on evm_log_poller_blocks  (cost=0.56..1620667.95 rows=10188000 width=62) (actual time=0.158..0.158 rows=1 loops=1)
         Index Cond: (evm_chain_id = '5'::numeric)
 Planning Time: 0.422 ms
 Execution Time: 0.233 ms
```
</details>

### Index added to the evm_log

<details>
   <summary> Explain analyze test - 6.5s -> 0.2ms </summary>

Before
```
Sort  (cost=1094581.76..1094581.76 rows=1 width=231) (actual time=6499.332..6499.332 rows=0 loops=1)
   Sort Key: created_at
   Sort Method: quicksort  Memory: 25kB
   ->  Gather  (cost=1000.00..1094581.75 rows=1 width=231) (actual time=6499.303..6501.945 rows=0 loops=1)
         Workers Planned: 2
         Workers Launched: 2
         ->  Parallel Seq Scan on evm_logs  (cost=0.00..1093581.65 rows=1 width=231) (actual time=6492.517..6492.517 rows=0 loops=3)
               Filter: ((created_at > '2023-05-31 07:29:11.29+00'::timestamp with time zone) AND (evm_chain_id = '420'::numeric) AND (address = '\xAAAA'::bytea) AND (event_sig = '\xBBBB'::bytea))
               Rows Removed by Filter: 6666667
 Planning Time: 0.583 ms
 Execution Time: 6502.046 ms
```

After
```
Index Scan using idx_evm_logs_ordered_by_block_and_created_at on evm_logs  (cost=0.69..8.71 rows=1 width=231) (actual time=0.086..0.086 rows=0 loops=1)
   Index Cond: ((evm_chain_id = '420'::numeric) AND (address = '\xAAAA'::bytea) AND (event_sig = '\xBBBB'::bytea) AND (created_at > '2023-05-31 07:29:11.29+00'::timestamp with time zone))
 Planning Time: 0.422 ms
 Execution Time: 0.141 ms
(4 rows)
```
</details>